+++ server/api/api/v1/accounts/id/index.ts
@@ -1,0 +1,51 @@
+import { applyConfig, auth, handleZodError } from "@/api";
+import { errorResponse, jsonResponse } from "@/response";
+import type { Hono } from "@hono/hono";
+import { zValidator } from "@hono/zod-validator";
+import { and, eq, isNull } from "drizzle-orm";
+import { z } from "zod";
+import { RolePermissions, Users } from "~/drizzle/schema";
+import { User } from "~/packages/database-interface/user";
+
+export const meta = applyConfig({
+    allowedMethods: ["GET"],
+    ratelimits: {
+        max: 30,
+        duration: 60,
+    },
+    route: "/api/v1/accounts/id",
+    auth: {
+        required: false,
+        oauthPermissions: [],
+    },
+    permissions: {
+        required: [RolePermissions.Search],
+    },
+});
+
+export const schemas = {
+    query: z.object({
+        username: z.string().min(1).max(512),
+    }),
+};
+
+export default (app: Hono) =>
+    app.on(
+        meta.allowedMethods,
+        meta.route,
+        zValidator("query", schemas.query, handleZodError),
+        auth(meta.auth, meta.permissions),
+        async (context) => {
+            const { username } = context.req.valid("query");
+
+            const user = await User.fromSql(
+                and(eq(Users.username, username), isNull(Users.instanceId)),
+            );
+
+            if (!user) {
+                return errorResponse("User not found", 404);
+            }
+
+            return jsonResponse(user.toApi());
+        },
+    );